fix: improve wait polling robustness

+++ cmd/remove.go
@@ -6,6 +6,10 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var (
+	waitForDelete bool
+)
+
 var removeCmd = &cobra.Command{
 	Use:   "rm <name>",
 	Short: "Remove job from cluster",
@@ -16,8 +20,28 @@ var removeCmd = &cobra.Command{
 		}
 
 		name := args[0]
+
+		job, err := kubectx.GetJob(name)
+		if err != nil {
+			return fmt.Errorf("unable to get job: %w", err)
+		}
+
+		if job == nil {
+			fmt.Printf("Nothing to delete: no job named %s found\n", name)
+			return nil
+		}
+
+		fmt.Printf("Deleting job %s...\n", name)
 		if err := kubectx.DeleteJob(name); err != nil {
 			return fmt.Errorf("unable to delete job: %w", err)
+		}
+
+		if !waitForDelete {
+			return nil
+		}
+
+		if err := waitUntilDeleted(name); err != nil {
+			return fmt.Errorf("timed out waiting for job to be deleted: %w", err)
 		}
 
 		return nil
@@ -26,4 +50,7 @@ var removeCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(removeCmd)
+
+	flags := removeCmd.Flags()
+	flags.BoolVarP(&waitForDelete, "wait", "w", false, "wait for job to be deleted")
 }
